Fix worker safety issues
