+++ js_modules/dagster-ui/packages/ui-core/src/app/Util.tsx
@@ -78,7 +78,7 @@ const msecFormatter = memoize((locale: string) => {
  * Return an i18n-formatted millisecond in seconds as a decimal, with no leading zero.
  */
 const formatMsecMantissa = (msec: number) =>
-  msecFormatter(navigator.language)
+  msecFormatter(typeof navigator !== 'undefined' ? navigator.language : 'en-US')
     .format(msec / 1000)
     .slice(-4);
 
@@ -110,23 +110,28 @@ export function breakOnUnderscores(str: string) {
 }
 
 export function patchCopyToRemoveZeroWidthUnderscores() {
-  document.addEventListener('copy', (event) => {
-    if (!event.clipboardData) {
-      // afaik this is always defined, but the TS field is optional
-      return;
-    }
+  if (typeof document !== 'undefined') {
+    document.addEventListener('copy', (event) => {
+      if (!event.clipboardData) {
+        // afaik this is always defined, but the TS field is optional
+        return;
+      }
 
-    // Note: This returns the text of the current selection if DOM
-    // nodes are selected. If the selection on the page is text within
-    // codemirror or an input or textarea, this returns "" and we fall
-    // through to the default pasteboard content.
-    const text = (window.getSelection() || '').toString().replace(/_\u200b/g, '_');
+      // Note: This returns the text of the current selection if DOM
+      // nodes are selected. If the selection on the page is text within
+      // codemirror or an input or textarea, this returns "" and we fall
+      // through to the default pasteboard content.
+      const text =
+        (typeof window !== 'undefined' ? window.getSelection() : '')
+          ?.toString()
+          .replace(/_\u200b/g, '_') || '';
 
-    if (text.length) {
-      event.preventDefault();
-      event.clipboardData.setData('Text', text);
-    }
-  });
+      if (text.length) {
+        event.preventDefault();
+        event.clipboardData.setData('Text', text);
+      }
+    });
+  }
 }
 
 export function asyncMemoize<T, R, U extends (arg: T, ...rest: any[]) => PromiseLike<R>>(
@@ -254,4 +259,7 @@ export const gqlTypePredicate =
     return node.__typename === typename;
   };
 
-export const COMMON_COLLATOR = new Intl.Collator(navigator.language, {sensitivity: 'base'});
+export const COMMON_COLLATOR = new Intl.Collator(
+  typeof navigator !== 'undefined' ? navigator.language : 'en-US',
+  {sensitivity: 'base'},
+);
