+++ js_modules/dagster-ui/packages/ui-core/src/app/__tests__/Flags.test.tsx
@@ -81,7 +81,8 @@ describe('Feature Flags with In-Memory Cache and BroadcastChannel', () => {
 
   it('should react to changes in feature flags across contexts', () => {
     jest.isolateModules(() => {
-      const {useFeatureFlags, setFeatureFlags} = require('../Flags');
+      const {setFeatureFlags} = require('../Flags');
+      const {useFeatureFlags} = require('../useFeatureFlags');
       const {result} = renderHook(() => useFeatureFlags());
 
       expect(result.current[FeatureFlag.__TestFlagDefaultNone]).toBe(false); // Default
@@ -103,7 +104,8 @@ describe('Feature Flags with In-Memory Cache and BroadcastChannel', () => {
 
   it('should handle unset state correctly', () => {
     jest.isolateModules(() => {
-      const {useFeatureFlags, setFeatureFlags} = require('../Flags');
+      const {setFeatureFlags} = require('../Flags');
+      const {useFeatureFlags} = require('../useFeatureFlags');
       const {result} = renderHook(() => useFeatureFlags());
 
       expect(result.current[FeatureFlag.__TestFlagDefaultNone]).toBe(false); // Default
