+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/AssetNodeFacetsUtil.ts
@@ -1,0 +1,11 @@
+export enum AssetNodeFacet {
+  UnsyncedTag = 'unsynced-tag',
+  Description = 'description',
+  Owner = 'owner',
+  LatestEvent = 'latest-event',
+  Checks = 'checks',
+  Freshness = 'freshness',
+  Status = 'status',
+  KindTag = 'kind-tag',
+  Automation = 'automation',
+}
