+++ js_modules/dagster-ui/packages/ui-core/src/hooks/useStateWithStorage.tsx
@@ -1,26 +1,6 @@
 import * as React from 'react';
 
-export function getJSONForKey(key: string) {
-  let stored = undefined;
-  try {
-    if (typeof window === 'undefined') {
-      stored = self.localStorage.getItem(key);
-    } else {
-      stored = window.localStorage.getItem(key);
-    }
-    if (stored) {
-      return JSON.parse(stored);
-    }
-  } catch {
-    if (typeof stored === 'string') {
-      // With useStateWithStorage, some values like timezone are moving from `UTC` to `"UTC"`
-      // in LocalStorage. To read the old values, pass through raw string values. We can
-      // remove this a few months after 0.14.1 is released.
-      return stored;
-    }
-    return undefined;
-  }
-}
+import {getJSONForKey} from '../util/getJSONForKey';
 
 const DID_WRITE_LOCALSTORAGE = '';
